backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# LangChain ecosystem
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# Data validation
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0

# Supabase
supabase>=2.0.0
